security implementado

+++ PROYECTO/src/main/java/org/example/proyecto/Item/Domain/ItemService.java
@@ -7,9 +7,13 @@ import org.example.proyecto.Item.dto.ItemRequestDto;
 import org.example.proyecto.Item.dto.ItemResponseDto;
 import org.example.proyecto.Usuario.Domain.Usuario;
 import org.example.proyecto.Usuario.infrastructure.UsuarioRepository;
+import org.example.proyecto.auth.utils.AuthorizationUtils;
 import org.example.proyecto.exception.ResourceNotFoundException;
+import org.example.proyecto.exception.UnauthorizeOperationException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import java.util.stream.Collectors;
 
@@ -20,15 +24,17 @@ public class ItemService {
     private final ItemRepository itemRepository;
     private final CategoryRepository categoryRepository;
     private final UsuarioRepository usuarioRepository;
+    private final AuthorizationUtils authorizationUtils;
 
     @Autowired
     ModelMapper modelMapper;
 
     @Autowired
-    public ItemService(ItemRepository itemRepository, CategoryRepository categoryRepository, UsuarioRepository usuarioRepository) {
+    public ItemService(ItemRepository itemRepository, CategoryRepository categoryRepository, UsuarioRepository usuarioRepository, AuthorizationUtils authorizationUtils) {
         this.itemRepository = itemRepository;
         this.categoryRepository = categoryRepository;
         this.usuarioRepository = usuarioRepository;
+        this.authorizationUtils = authorizationUtils;
     }
 
     public ItemResponseDto createItem(ItemRequestDto itemDto) {
@@ -37,6 +43,7 @@ public class ItemService {
 
         Usuario user = usuarioRepository.findById(itemDto.getUser_id())
                 .orElseThrow(() -> new ResourceNotFoundException("Usuario no encontrado"));
+
 
         //mapeamos de itemDto a item
         Item item = modelMapper.map(itemDto, Item.class);
@@ -68,6 +75,9 @@ public class ItemService {
         Usuario usuario = usuarioRepository.findById(itemRequestDto.getUser_id())
                 .orElseThrow(() -> new ResourceNotFoundException("Usuario no encontrado"));
 
+        if (!authorizationUtils.isAdminOrResourceOwner(existingItem.getUsuario().getId())) {
+            throw new UnauthorizeOperationException("You do not have permission to update this item.");
+        }
 
         modelMapper.map(itemRequestDto, existingItem);
 
@@ -100,7 +110,12 @@ public class ItemService {
 
     public void deleteItem(Long itemId) {
         Item item = itemRepository.findById(itemId)
-                .orElseThrow(() -> new ResourceNotFoundException("Item no encontrado"));
+                .orElseThrow(() -> new ResourceNotFoundException("Item not found"));
+
+        if (!authorizationUtils.isAdminOrResourceOwner(item.getUsuario().getId())) {
+            throw new UnauthorizeOperationException("You do not have permission to delete this item.");
+        }
+
         itemRepository.delete(item);
     }
 
@@ -121,6 +136,8 @@ public class ItemService {
     public List<ItemResponseDto> getItemsByCategory(Long categoryId) {
         List<Item> items = itemRepository.findByCategoryId(categoryId);
 
+
+
         return items.stream()
                 .map(item -> {
                     ItemResponseDto responseDto = modelMapper.map(item, ItemResponseDto.class);
@@ -139,6 +156,35 @@ public class ItemService {
                     ItemResponseDto responseDto = modelMapper.map(item, ItemResponseDto.class);
                     responseDto.setCategoryName(item.getCategory().getName());  // Seteamos el nombre de la categoría
                     responseDto.setUserName(item.getUsuario().getEmail());  // Seteamos el nombre (o email) del usuario
+                    return responseDto;
+                })
+                .collect(Collectors.toList());
+    }
+
+    public List<ItemResponseDto> getUserItems() {
+        // Obtener el email del usuario autenticado usando Spring Security
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        String email;
+        if (principal instanceof UserDetails) {
+            email = ((UserDetails) principal).getUsername();
+        } else {
+            throw new UnauthorizeOperationException("Usuario no autenticado");
+        }
+
+        // Buscar el usuario por su email
+        Usuario usuario = usuarioRepository.findByEmail(email)
+                .orElseThrow(() -> new ResourceNotFoundException("Usuario no encontrado"));
+
+        // Obtener los ítems del usuario autenticado
+        List<Item> items = itemRepository.findByUsuarioId(usuario.getId());
+
+        // Mapear los ítems a ItemResponseDto
+        return items.stream()
+                .map(item -> {
+                    ItemResponseDto responseDto = modelMapper.map(item, ItemResponseDto.class);
+                    responseDto.setCategoryName(item.getCategory().getName());
+                    responseDto.setUserName(item.getUsuario().getEmail());
                     return responseDto;
                 })
                 .collect(Collectors.toList());
